Cleanup after merge
