+++ api/src/services/files/lib/decompress-response.ts
@@ -1,0 +1,75 @@
+import type { AxiosResponse } from 'axios';
+import type { Readable } from 'node:stream';
+import { PassThrough, Transform } from 'node:stream';
+import zlib from 'node:zlib';
+
+export function decompressResponse(stream: Readable, headers: AxiosResponse['headers']) {
+	const contentEncoding = (headers['content-encoding'] || '').toLowerCase();
+
+	if (!['gzip', 'deflate', 'br'].includes(contentEncoding)) {
+		return stream;
+	}
+
+	let isEmpty = true;
+
+	const checker = new Transform({
+		transform(data, _encoding, callback) {
+			if (isEmpty === false) {
+				callback(null, data);
+				return;
+			}
+
+			isEmpty = false;
+
+			handleContentEncoding(data);
+
+			callback(null, data);
+		},
+
+		flush(callback) {
+			callback();
+		},
+	});
+
+	const finalStream = new PassThrough({
+		autoDestroy: false,
+		destroy(error, callback) {
+			stream.destroy();
+
+			callback(error);
+		},
+	});
+
+	stream.pipe(checker);
+
+	return finalStream;
+
+	function handleContentEncoding(data: any) {
+		let decompressStream;
+
+		if (contentEncoding === 'br') {
+			decompressStream = zlib.createBrotliDecompress();
+		} else if (contentEncoding === 'deflate' && isDeflateAlgorithm(data)) {
+			decompressStream = zlib.createInflateRaw();
+		} else {
+			decompressStream = zlib.createUnzip();
+		}
+
+		decompressStream.once('error', (error) => {
+			if (isEmpty && !stream.readable) {
+				finalStream.end();
+				return;
+			}
+
+			finalStream.destroy(error);
+		});
+
+		checker.pipe(decompressStream).pipe(finalStream);
+	}
+
+	function isDeflateAlgorithm(data: any) {
+		const DEFLATE_ALGORITHM_HEADER = 0x08;
+
+		return data.length > 0 && (data[0] & DEFLATE_ALGORITHM_HEADER) === 0;
+	}
+}
